Update installation docs
The installation docs have a simple process like in other gems (e.g. bootsnap) with an Envfile starter. Last step is a link to Types section to start ENV configuration.

Includes a helpful section on how to do a pre-deploy env check. The main requirement is the host app has to provide its current env in a parsable format (this can vary) so a Ruby script can load it using `ENV.replace(hash)`.