Replace CGI usage with URI for Ruby 3.5 compatibility (#7)
* Replace CGI with URI for Ruby 3.5 support

In Ruby 3.5, most of the cgi gem is removed which includes `CGI.parse`. We can replace this with `URI.decode_www_form` but it has some slight differences. Seems any blank value now always comes back as nil rather nil or an empty string depending so this can be a breaking change.

https://github.com/rouge-ruby/rouge/pull/2131

* Update CHANGELOG.md